+++ src/interface/swarm_default_names.hpp
@@ -19,11 +19,10 @@
 #include "swarm_pack.hpp"
 
 #define SWARM_VARIABLE(type, ns, varname)                                                \
-  struct varname : public parthenon::swarm_variable_names::base_t<false, type> {         \
+  struct varname : public parthenon::swarm_variable_names::base_t<type> {                \
     template <class... Ts>                                                               \
     KOKKOS_INLINE_FUNCTION varname(Ts &&...args)                                         \
-        : parthenon::swarm_variable_names::base_t<false, type>(                          \
-              std::forward<Ts>(args)...) {}                                              \
+        : parthenon::swarm_variable_names::base_t<type>(std::forward<Ts>(args)...) {}    \
     static std::string name() { return #ns "." #varname; }                               \
   }
